24.08.02 merge job

+++ include/Rendering.h
@@ -11,10 +11,10 @@ public:
     std::string* real_color1;
     std::string* real_depth1;
     std::vector<double> lookAtVector = {0.000588, 0.739846, -0.903124, 0.087284, 1.51642, -0.279094, 0.163545, -0.628984, 0.760021};
-    float fx = 386.52199190267083;//焦距(x轴上)
-    float fy = 387.32300428823663;//焦距(y轴上)
-    float cx = 326.5103569741365;//图像中心点(x轴)
-    float cy = 237.40293732598795;//图像中心点(y轴)
+    double fx = 386.52199190267083;//焦距(x轴上)
+    double fy = 387.32300428823663;//焦距(y轴上)
+    double cx = 326.5103569741365;//图像中心点(x轴)
+    double cy = 237.40293732598795;//图像中心点(y轴)
     std::vector<vsg::dmat4> model_transforms;
     std::vector<vsg::dmat4> init_model_transforms;
     std::vector<std::string> model_paths;
